refactor: move iacc to wasm to save on data copies

+++ src/math/wasm.ts
@@ -1,7 +1,7 @@
 import { Complex } from '@iamsquare/complex.js';
 
-export function complexObjectFormIntoFlatForm(array: Complex[]): Float64Array {
-  const flattened = new Float64Array(array.length * 2);
+export function complexObjectFormIntoFlatForm(array: Complex[]): Float32Array {
+  const flattened = new Float32Array(array.length * 2);
 
   for (let i = 0; i < array.length; i += 1) {
     flattened[i * 2] = array[i].getRe();
@@ -11,7 +11,7 @@ export function complexObjectFormIntoFlatForm(array: Complex[]): Float64Array {
   return flattened;
 }
 
-export function complexFlatFormIntoObjectForm(array: Float64Array): Complex[] {
+export function complexFlatFormIntoObjectForm(array: Float32Array): Complex[] {
   const objectForm: Complex[] = [];
 
   for (let i = 0; i < array.length; i += 2) {
